warning fixes
git-svn-id: http://svn.miranda-ng.org/main/trunk@13728 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/MSN/src/msn_commands.cpp
@@ -732,7 +732,7 @@ void CMsnProto::MSN_ProcessStatusMessage(ezxml_t xmli, const char* wlid)
 #endif
 }
 
-void CMsnProto::MSN_ProcessNotificationMessage(char* buf, unsigned len)
+void CMsnProto::MSN_ProcessNotificationMessage(char* buf, size_t len)
 {
 	if (buf == NULL) return;
 	ezxml_t xmlnot = ezxml_parse_str(buf, len);
@@ -1047,7 +1047,7 @@ LBL_InvalidCommand:
 								MEVENT hDbEvent;
 								bool bDuplicate = false;
 								DBEVENTINFO dbei = { sizeof(dbei) };
-								DWORD cbBlob = strlen(message);
+								DWORD cbBlob = (DWORD)strlen(message);
 								dbei.cbBlob = cbBlob;
 								BYTE *pszMsgBuf = (BYTE*)mir_calloc(cbBlob);
 								if (pszMsgBuf) {
